requirements.txt
pytest
selenium
python-dotenv
pytest-xdist
allure-pytest
webdriver-manager